requirements.txt
python-telegram-bot==20.3
cohere==3.4.0
